feat: implement order management controller and add chatbot functionality for users

+++ app/Livewire/Chatbot.php
@@ -181,7 +181,7 @@ class Chatbot extends Component
             $order = Order::find($routeOrder);
         }
 
-        if (!$order || $order->user_id !== Auth::id()) {
+        if (!$order || (int) $order->user_id !== (int) Auth::id()) {
             return;
         }
